+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:meedu/router.dart' show MNavigator;
+import 'package:meedu/router.dart' as router;
 import 'package:meedu/get.dart';
 import 'modules/home/home_page.dart';
 import 'modules/numbers/numbers_page.dart';
@@ -14,7 +14,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      navigatorKey: MNavigator.instance.navigatorKey,
+      navigatorKey: router.navigatorKey,
       theme: ThemeData(
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,