+++ lib/src/router/navigator.dart
@@ -1,13 +1,12 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
-class MNavigator {
-  MNavigator._internal();
-  static MNavigator _instance = MNavigator._internal();
-  static MNavigator get instance => _instance;
+class Navigator {
+  Navigator._internal();
+  static Navigator _instance = Navigator._internal();
+  static Navigator get instance => _instance;
 
   final GlobalKey<NavigatorState> _navigatorKey =
       new GlobalKey<NavigatorState>();
-
   GlobalKey<NavigatorState> get navigatorKey => _navigatorKey;
   NavigatorState get state => _navigatorKey.currentState;
 }