+++ lib/src/router/router.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:meedu/meedu.dart';
+import 'navigator.dart' as navigator;
+
+GlobalKey<NavigatorState> get navigatorKey =>
+    navigator.Navigator.instance.navigatorKey;
 
 /// Push the given [page] onto the navigator.
 Future<T> push<T>(Widget page, {Object arguments}) {
@@ -11,7 +14,7 @@ Future<T> push<T>(Widget page, {Object arguments}) {
       arguments: arguments,
     ),
   );
-  return MNavigator.instance.state.push(route);
+  return navigator.Navigator.instance.state.push(route);
 }
 
 /// Replace the current [page] of the navigator by pushing the given [page] and then
@@ -25,24 +28,24 @@ Future<T> pushReplacement<T>(Widget page, {Object arguments}) {
       arguments: arguments,
     ),
   );
-  return MNavigator.instance.state.pushReplacement(route);
+  return navigator.Navigator.instance.state.pushReplacement(route);
 }
 
 /// Push a named route onto the navigator.
 Future<T> pushNamed<T>(String routeName, {Object arguments}) {
-  return MNavigator.instance.state
+  return navigator.Navigator.instance.state
       .pushNamed<T>(routeName, arguments: arguments);
 }
 
 /// Pop the current route off the navigator and push a named route in its place
 Future<T> popAndPushNamed<T>(String routeName, {Object arguments}) {
-  return MNavigator.instance.state
+  return navigator.Navigator.instance.state
       .popAndPushNamed(routeName, arguments: arguments);
 }
 
 /// replace the current page with a new route name
 Future<T> pushReplacementNamed<T>(String routeName, {Object arguments}) {
-  return MNavigator.instance.state
+  return navigator.Navigator.instance.state
       .pushReplacementNamed(routeName, arguments: arguments);
 }
 
@@ -52,18 +55,18 @@ Future<T> pushNamedAndRemoveUntil<T>(
   bool Function(Route<dynamic>) predicate,
   Object arguments,
 }) {
-  return MNavigator.instance.state.pushNamedAndRemoveUntil(
+  return navigator.Navigator.instance.state.pushNamedAndRemoveUntil(
       routeName, predicate ?? () => false,
       arguments: arguments);
 }
 
 /// remove all pages in the stack until [predicate]
 void popUntil([bool Function(Route<dynamic>) predicate]) {
-  MNavigator.instance.state.popUntil(predicate ?? () => false);
+  navigator.Navigator.instance.state.popUntil(predicate ?? () => false);
 }
 
 /// return true if we can do pop
-bool canPop() => MNavigator.instance.state.canPop();
+bool canPop() => navigator.Navigator.instance.state.canPop();
 
 /// return the arguments of the current page
 Object arguments(BuildContext context) {