+++ lib/state.dart
@@ -1,5 +1,5 @@
 library state;
 
-export 'src/state_manager/meedu_builder.dart';
-export 'src/state_manager/meedu_controller.dart';
-export 'src/state_manager/meedu_provider.dart';
+export 'src/state_managment/meedu_builder.dart';
+export 'src/state_managment/meedu_controller.dart';
+export 'src/state_managment/meedu_provider.dart';